moving background and other style updates

+++ deploy.sh
@@ -9,12 +9,18 @@ npm install
 # Build the React app
 npm run build
 
+# Check if build was successful
+if [ ! -d "build" ]; then
+    echo "Build failed. Exiting."
+    exit 1
+fi
+
 # Move build files to the public_html directory
 mv build/* .
 mv build/.* . 2>/dev/null
 rmdir build
 
-# Optionally, remove development files
-rm -rf src public
+# Optionally, remove source files to save space
+rm -rf src
 
 echo "Deployment completed"
